backend-microservices/users-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-multipart==0.0.12
PyJWT==2.9.0
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
